update the calendar component
Hard coded the complete function so that the dash would appear after reloading the task. But still unsure of how to do that with the edit function. I suspect the agenda item other than the first 1 not updating is that there is a problem with the importede agenda component from react native calendar.
https://github.com/wix/react-native-calendars/issues/1855